Title: 서로다른 문자열의 개수, Time: 113ms, Memory: 8MB, Status: Passed - CodeTree

+++ 240806/서로다른 문자열의 개수/number-of-different-strings.java
@@ -5,7 +5,7 @@ public class Main {
     public static void main(String[] args) throws Exception {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
-        int[][] dp = new int[n+1][n+1];
+        long[][] dp = new long[n+1][n+1];
 
         // b의 개수 >= a의 개수
         for (int b=1;b<=n;b++) {
